OperationsNPI/requirements.txt
pymysql
fastapi
uvicorn
sqlalchemy
python-dotenv
pytest-asyncio

